Do not infect via symlinks
Changed behavior if the virus cannot open the victim as read-write:
    * Give up if the file is a directory.
    * Otherwise, open it again as read-only but as a hard-link.
    * Give up if it still cannot open the file.

If the path refers to a symlink, then replacing it with a new file
does not actually infect the original file.
Also, "infecting" a symlink is too obvious,
since the replacement file would must be of a different type.